+++ src/components/NostoSkuOptions.ts
@@ -2,6 +2,7 @@ import { assertRequired, intersectionOf } from "@/utils"
 import { injectStore, Store } from "./NostoProduct/store"
 import { customElement } from "./decorators"
 import { syncSkuData } from "./common"
+import { NostoElement } from "./NostoElement"
 
 /**
  * A custom element that manages SKU (Stock Keeping Unit) options in a product selection interface.
@@ -29,7 +30,7 @@ import { syncSkuData } from "./common"
  * ```
  */
 @customElement("nosto-sku-options")
-export class NostoSkuOptions extends HTMLElement {
+export class NostoSkuOptions extends NostoElement {
   static attributes = {
     name: String
   }