+++ src/components/decorators.ts
@@ -1,4 +1,10 @@
-type FieldType<T> = T extends string ? StringConstructor : T extends boolean ? BooleanConstructor : never
+type FieldType<T> = T extends string
+  ? StringConstructor
+  : T extends number
+    ? NumberConstructor
+    : T extends boolean
+      ? BooleanConstructor
+      : never
 
 type ConstructorMetadata<T extends HTMLElement> = {
   new (): T
@@ -9,15 +15,24 @@ export function customElement<T extends HTMLElement>(tagName: string) {
   return function (constructor: ConstructorMetadata<T>) {
     if (constructor.attributes) {
       Object.entries(constructor.attributes).forEach(([fieldName, type]) => {
-        const attribute = toKebabCase(fieldName)
-        const property = (type === String ? stringAttribute : booleanAttribute)(attribute)
-        Object.defineProperty(constructor.prototype, fieldName, property)
+        const propertyDescriptor = getPropertyDescriptor(fieldName, type)
+        Object.defineProperty(constructor.prototype, fieldName, propertyDescriptor)
       })
     }
     if (!window.customElements.get(tagName)) {
       window.customElements.define(tagName, constructor)
     }
   }
+}
+
+function getPropertyDescriptor(propertyName: string, type: unknown) {
+  const attributeName = toKebabCase(propertyName)
+  if (type === Boolean) {
+    return booleanAttribute(attributeName)
+  } else if (type === Number) {
+    return numberAttribute(attributeName)
+  }
+  return stringAttribute(attributeName)
 }
 
 function toKebabCase(str: string) {
@@ -29,7 +44,7 @@ function stringAttribute(attributeName: string) {
     get(this: HTMLElement) {
       return this.getAttribute(attributeName)!
     },
-    set(this: HTMLElement, value: string) {
+    set(this: HTMLElement, value?: string) {
       if (value === null || value === undefined) {
         this.removeAttribute(attributeName)
       } else {
@@ -48,6 +63,24 @@ function booleanAttribute(attributeName: string) {
     },
     set(this: HTMLElement, value: boolean) {
       this.toggleAttribute(attributeName, value)
+    },
+    configurable: true,
+    enumerable: true
+  }
+}
+
+function numberAttribute(attributeName: string) {
+  return {
+    get(this: HTMLElement) {
+      const value = this.getAttribute(attributeName)
+      return value ? Number(value) : undefined
+    },
+    set(this: HTMLElement, value?: number) {
+      if (value === null || value === undefined) {
+        this.removeAttribute(attributeName)
+      } else {
+        this.setAttribute(attributeName, value.toString())
+      }
     },
     configurable: true,
     enumerable: true